semana 10 y porfin funciono :)

+++ src/main/java/com/tiendaTech/tienda/controller/CategoriaController.java
@@ -15,49 +15,60 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 @Controller
 @RequestMapping("/categoria")
 public class CategoriaController {
-    private final CategoriaService categoriaService;
+    
+     private final CategoriaService categoriaService;
     private final MessageSource messageSource;
+
     public CategoriaController(CategoriaService categoriaService, MessageSource messageSource) {
         this.categoriaService = categoriaService;
         this.messageSource = messageSource;
     }
+
     @GetMapping("/listado")
     public String listado(Model model) {
+        
         var categorias = categoriaService.getCategorias(false);
+        
         model.addAttribute("categorias", categorias);
-        model.addAttribute("totalCategorias", categorias.size()); // ✅ ahora List tiene .size()
+        model.addAttribute("totalCategorias", categorias.size());
         return "/categoria/listado";
     }
+
     @PostMapping("/guardar")
     public String guardar(@Valid Categoria categoria, @RequestParam MultipartFile imagenFile, RedirectAttributes redirectAttributes) {
+
         categoriaService.save(categoria, imagenFile);
         redirectAttributes.addFlashAttribute("todoOk", messageSource.getMessage("mensaje.actualizado", null, Locale.getDefault()));
+
         return "redirect:/categoria/listado";
     }
+
     @PostMapping("/eliminar")
-    public String eliminar(@RequestParam Long idCategoria, RedirectAttributes redirectAttributes) { // ✅ CAMBIO: Integer → Long
+    public String eliminar(@RequestParam Integer idCategoria, RedirectAttributes redirectAttributes) {
         String titulo = "todoOk";
         String detalle = "mensaje.eliminado";
         try {
             categoriaService.delete(idCategoria);
         } catch (IllegalArgumentException e) {
-            titulo = "error";
+            titulo = "error"; // Captura la excepción de argumento inválido para el mensaje de "no existe"
             detalle = "categoria.error01";
         } catch (IllegalStateException e) {
-            titulo = "error";
+            titulo = "error"; // Captura la excepción de estado ilegal para el mensaje de "datos asociados"
             detalle = "categoria.error02";
         } catch (Exception e) {
-            titulo = "error";
+            titulo = "error";  // Captura cualquier otra excepción inesperada
             detalle = "categoria.error03";
         }
         redirectAttributes.addFlashAttribute(titulo, messageSource.getMessage(detalle, null, Locale.getDefault()));
         return "redirect:/categoria/listado";
     }
+
     @GetMapping("/modificar/{idCategoria}")
-    public String modificar(@PathVariable("idCategoria") Long idCategoria, Model model, RedirectAttributes redirectAttributes) { // ✅ CAMBIO: Integer → Long
+    public String modificar(@PathVariable("idCategoria") Integer idCategoria, Model model, RedirectAttributes redirectAttributes) {
         Optional<Categoria> categoriaOpt = categoriaService.getCategoria(idCategoria);
         if (categoriaOpt.isEmpty()) {
             redirectAttributes.addFlashAttribute("error", messageSource.getMessage("categoria.error01", null, Locale.getDefault()));
